Moved DeskClientUtils specific test methods into DeskClientUtilsTest
