+++ api-client/src/test/java/com/desk/java/apiclient/util/DeskClientUtilsTest.java
@@ -1,0 +1,64 @@
+/*
+ * Copyright (c) 2015, Salesforce.com, Inc.
+ * All rights reserved.
+ *
+ * Redistribution and use in source and binary forms, with or without modification, are permitted provided
+ * that the following conditions are met:
+ *
+ *    Redistributions of source code must retain the above copyright notice, this list of conditions and the
+ *    following disclaimer.
+ *
+ *    Redistributions in binary form must reproduce the above copyright notice, this list of conditions and
+ *    the following disclaimer in the documentation and/or other materials provided with the distribution.
+ *
+ *    Neither the name of Salesforce.com, Inc. nor the names of its contributors may be used to endorse or
+ *    promote products derived from this software without specific prior written permission.
+ *
+ * THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS" AND ANY EXPRESS OR IMPLIED
+ * WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A
+ * PARTICULAR PURPOSE ARE DISCLAIMED. IN NO EVENT SHALL THE COPYRIGHT OWNER OR CONTRIBUTORS BE LIABLE FOR
+ * ANY DIRECT, INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING, BUT NOT LIMITED
+ * TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE, DATA, OR PROFITS; OR BUSINESS INTERRUPTION)
+ * HOWEVER CAUSED AND ON ANY THEORY OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
+ * NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE, EVEN IF ADVISED OF THE
+ * POSSIBILITY OF SUCH DAMAGE.
+ */
+
+package com.desk.java.apiclient.util;
+
+import org.junit.Test;
+
+import static com.desk.java.apiclient.DeskClient.OAUTH_ACCESS_URL;
+import static com.desk.java.apiclient.DeskClient.OAUTH_AUTHORIZE_URL;
+import static com.desk.java.apiclient.DeskClient.OAUTH_REQUEST_URL;
+import static com.desk.java.apiclient.DeskClient.PROTOCOL_CONNECT;
+import static org.junit.Assert.assertEquals;
+
+/**
+ * Unit tests for {@link DeskClientUtils}.
+ *
+ * Created by Jerrell Mardis
+ * Copyright (c) 2015 Desk.com. All rights reserved.
+ */
+public class DeskClientUtilsTest {
+
+    private static final String TEST_HOST_NAME = "test.test.com";
+
+    @Test
+    public void oAuthRequestUrlReturnsCorrectUrl() throws Exception {
+        final String requestUrl = DeskClientUtils.oAuthRequestUrl(TEST_HOST_NAME);
+        assertEquals(PROTOCOL_CONNECT + TEST_HOST_NAME + OAUTH_REQUEST_URL, requestUrl);
+    }
+
+    @Test
+    public void oAuthAccessUrlReturnsCorrectUrl() throws Exception {
+        final String accessUrl = DeskClientUtils.oAuthAccessUrl(TEST_HOST_NAME);
+        assertEquals(PROTOCOL_CONNECT + TEST_HOST_NAME + OAUTH_ACCESS_URL, accessUrl);
+    }
+
+    @Test
+    public void oAuthAuthorizeUrlReturnsCorrectUrl() throws Exception {
+        final String authorizeUrl = DeskClientUtils.oAuthAuthorizeUrl(TEST_HOST_NAME);
+        assertEquals(PROTOCOL_CONNECT + TEST_HOST_NAME + OAUTH_AUTHORIZE_URL, authorizeUrl);
+    }
+}
